testing: quick refactor
